+++ src/admin/features/authentication/components/wx-login.tsx
@@ -22,7 +22,10 @@ const WxLogin: FC<Props> = (props: Props) => {
     }
 
     const wxLoginMock = () => {
-        window.location.href = "/api/weixin/login?wxid=111a"
+        const query = new URLSearchParams();
+        query.append('wxid', '');
+        query.append('state', props.useType);
+        window.location.href = "/api/weixin/login?" + query.toString()
     }
